LPD-28845 Build REST
