+++ modules/apps/headless/headless-admin-site/headless-admin-site-impl/src/main/java/com/liferay/headless/admin/site/internal/graphql/servlet/v1_0/ServletDataImpl.java
@@ -254,11 +254,6 @@ public class ServletDataImpl implements ServletData {
 							FragmentCompositionResourceImpl.class,
 							"putSiteSiteByExternalReferenceCodeFragmentComposition"));
 					put(
-						"mutation#createSiteSiteByExternalReferenceCodePageElementFragmentComposition",
-						new ObjectValuePair<>(
-							FragmentCompositionResourceImpl.class,
-							"postSiteSiteByExternalReferenceCodePageElementFragmentComposition"));
-					put(
 						"mutation#createSiteSiteByExternalReferenceCodeMasterPage",
 						new ObjectValuePair<>(
 							MasterPageResourceImpl.class,
@@ -308,6 +303,11 @@ public class ServletDataImpl implements ServletData {
 						new ObjectValuePair<>(
 							PageElementResourceImpl.class,
 							"putSiteSiteByExternalReferenceCodePageElement"));
+					put(
+						"mutation#createSiteSiteByExternalReferenceCodePageElementFragmentComposition",
+						new ObjectValuePair<>(
+							PageElementResourceImpl.class,
+							"postSiteSiteByExternalReferenceCodePageElementFragmentComposition"));
 					put(
 						"mutation#createSiteSiteByExternalReferenceCodePageExperiencePageElement",
 						new ObjectValuePair<>(
