+++ modules/apps/headless/headless-admin-site/headless-admin-site-test/src/testIntegration/java/com/liferay/headless/admin/site/resource/v1_0/test/BasePageElementResourceTestCase.java
@@ -216,6 +216,29 @@ public abstract class BasePageElementResourceTestCase {
 	}
 
 	@Test
+	public void testPostSiteSiteByExternalReferenceCodePageElementFragmentComposition()
+		throws Exception {
+
+		PageElement randomPageElement = randomPageElement();
+
+		PageElement postPageElement =
+			testPostSiteSiteByExternalReferenceCodePageElementFragmentComposition_addPageElement(
+				randomPageElement);
+
+		assertEquals(randomPageElement, postPageElement);
+		assertValid(postPageElement);
+	}
+
+	protected PageElement
+			testPostSiteSiteByExternalReferenceCodePageElementFragmentComposition_addPageElement(
+				PageElement pageElement)
+		throws Exception {
+
+		throw new UnsupportedOperationException(
+			"This method needs to be implemented");
+	}
+
+	@Test
 	public void testGetSiteSiteByExternalReferenceCodePageElementPageElementsPage()
 		throws Exception {
